+++ packages/shared/src/tests/fixtures.ts
@@ -160,7 +160,9 @@ export const ingressEgressEnvironment = ({
         Bitcoin: 0,
       },
     ),
-    boost_minimum_add_funds_amounts: createChainAssetMap(null),
+    boost_minimum_add_funds_amounts: createChainAssetMap('0x0', {
+      Btc: '0x186a0',
+    }),
   },
 });